refers #230 - Added TerminalSnapshot.getLogicalFields() in addition to getFields().
Purpose is to allows vendors (currentlyAPX) / OL to provide host fields
by attributes to be split to more by logical fields - 2 blanks or
more, different colors, etc.

It all started since Snapshot XML DTO messed up the trail XML, and
it tried to merge back logical field to fields by attributes, which it's
not his role. Moved this logic to
ApxTerminalSnapshot.initLogicalFields().

Other implementations still needs to implement logical fields (#169).
May be more useful for analyzer as well.

-Fixed fields building logic of H3270 which had many problems. Mostly
from the fact that it holds multi line fields, while OL break fields to
rows and can't allow this, and is 0 base, and holds fields for
attributes as well.

-Added partial test for H3270 trail